움직임에 감/가속 도입
